+++ TestSuite/yml.py
@@ -25,5 +25,11 @@ class YAML(File):
         yml_contents.update(update_obj)
         self.write_dict(yml_contents)
 
+    def delete_key(self, key: str):
+        yml_contents = self.read_dict()
+        if key in yml_contents:
+            del yml_contents[key]
+            self.write_dict(yml_contents)
+
     def update_description(self, description: str):
         self.update({'description': description})