+++ demisto_sdk/commands/common/constants.py
@@ -38,6 +38,7 @@ TESTS_DIR = 'Tests'
 DOC_FILES_DIR = 'doc_files'
 DOCUMENTATION_DIR = 'Documentation'
 JOBS_DIR = 'Jobs'
+PRE_PROCESS_RULES_DIR = 'PreProcessRules'
 
 SCRIPT = 'script'
 AUTOMATION = 'automation'
@@ -46,6 +47,7 @@ PLAYBOOK = 'playbook'
 TEST_PLAYBOOK = 'testplaybook'
 LAYOUT = 'layout'
 LAYOUTS_CONTAINER = 'layoutscontainer'
+PRE_PROCESS_RULES = 'pre-process-rules'
 INCIDENT_TYPE = 'incidenttype'
 INCIDENT_FIELD = 'incidentfield'
 INDICATOR_FIELD = 'indicatorfield'
@@ -93,6 +95,7 @@ class FileType(Enum):
     CONNECTION = 'canvas-context-connections'
     README = 'readme'
     RELEASE_NOTES = 'releasenotes'
+    RELEASE_NOTES_CONFIG = 'releasenotesconfig'
     DESCRIPTION = 'description'
     CHANGELOG = 'changelog'
     IMAGE = 'image'
@@ -112,6 +115,7 @@ class FileType(Enum):
     GENERIC_FIELD = 'genericfield'
     GENERIC_TYPE = 'generictype'
     GENERIC_DEFINITION = 'genericdefinition'
+    PRE_PROCESS_RULES = 'pre-process-rule'
     JOB = 'job'
 
 
@@ -133,6 +137,7 @@ RN_HEADER_BY_FILE_TYPE = {
     FileType.DASHBOARD: 'Dashboards',
     FileType.CONNECTION: 'Connections',
     FileType.MAPPER: 'Mappers',
+    FileType.PRE_PROCESS_RULES: 'PreProcess Rules',
     FileType.GENERIC_DEFINITION: 'Objects',
     FileType.GENERIC_MODULE: 'Modules',
     FileType.GENERIC_TYPE: 'Object Types',
@@ -158,6 +163,7 @@ ENTITY_TYPE_TO_DIR = {
     FileType.WIDGET.value: WIDGETS_DIR,
     FileType.BETA_INTEGRATION.value: INTEGRATIONS_DIR,
     FileType.MAPPER.value: CLASSIFIERS_DIR,
+    FileType.PRE_PROCESS_RULES.value: PRE_PROCESS_RULES_DIR,
     FileType.GENERIC_DEFINITION.value: GENERIC_DEFINITIONS_DIR,
     FileType.GENERIC_MODULE.value: GENERIC_MODULES_DIR,
     FileType.GENERIC_FIELD.value: GENERIC_FIELDS_DIR,
@@ -188,6 +194,7 @@ CONTENT_ENTITIES_DIRS = [
     GENERIC_TYPES_DIR,
     GENERIC_MODULES_DIR,
     GENERIC_DEFINITIONS_DIR,
+    PRE_PROCESS_RULES_DIR,
     JOBS_DIR
 ]
 
@@ -203,6 +210,8 @@ CONTENT_ENTITY_UPLOAD_ORDER = [
     CLASSIFIERS_DIR,
     WIDGETS_DIR,
     LAYOUTS_DIR,
+    DASHBOARDS_DIR,
+    PRE_PROCESS_RULES_DIR,
     DASHBOARDS_DIR,
     JOBS_DIR
 ]
@@ -407,6 +416,9 @@ PACKS_LAYOUT_JSON_REGEX = fr'{PACKS_LAYOUTS_DIR_REGEX}\/(?!layoutscontainer)([^/
 
 PACKS_LAYOUTS_CONTAINER_JSON_REGEX = fr'{PACKS_LAYOUTS_DIR_REGEX}\/layoutscontainer([^/]+)\.json'
 
+PACKS_PRE_PROCESS_RULES_DIR_REGEX = fr'{PACK_DIR_REGEX}/{PRE_PROCESS_RULES_DIR}'
+PACKS_PRE_PROCESS_RULES_JSON_REGEX = fr'{PACKS_PRE_PROCESS_RULES_DIR_REGEX}/(?:preprocessrule-)?([^/]+)\.json'
+
 PACKS_WIDGETS_DIR_REGEX = fr'{PACK_DIR_REGEX}\/{WIDGETS_DIR}'
 PACKS_WIDGET_JSON_REGEX = fr'{PACKS_WIDGETS_DIR_REGEX}\/([^/]+)\.json'
 
@@ -509,6 +521,7 @@ PACK_METADATA_USE_CASES = 'useCases'
 PACK_METADATA_KEYWORDS = 'keywords'
 PACK_METADATA_PRICE = 'price'
 PACK_METADATA_DEPENDENCIES = 'dependencies'
+PACK_METADATA_IRON_BANK_TAG = 'Iron Bank'
 
 PACK_METADATA_FIELDS = (PACK_METADATA_NAME, PACK_METADATA_DESC, PACK_METADATA_SUPPORT,
                         PACK_METADATA_CURR_VERSION, PACK_METADATA_AUTHOR, PACK_METADATA_URL, PACK_METADATA_CATEGORIES,
@@ -533,7 +546,8 @@ ID_IN_ROOT = [  # entities in which 'id' key is in the root
     'incident_type',
     'layoutscontainer',
     'mapper',
-    'job'```
+    'pre_process_rule',
+    'job'
 ]
 
 INTEGRATION_PREFIX = 'integration'
@@ -657,6 +671,10 @@ JSON_ALL_LAYOUTS_CONTAINER_REGEXES = [
     PACKS_LAYOUTS_CONTAINER_JSON_REGEX,
 ]
 
+JSON_ALL_PRE_PROCESS_RULES_REGEXES = [
+    PACKS_PRE_PROCESS_RULES_JSON_REGEX,
+]
+
 JSON_ALL_INCIDENT_FIELD_REGEXES = [
     PACKS_INCIDENT_FIELD_JSON_REGEX,
 ]
@@ -738,6 +756,7 @@ CHECKED_TYPES_REGEXES = [
     PACKS_INDICATOR_TYPE_JSON_REGEX,
     PACKS_LAYOUT_JSON_REGEX,
     PACKS_LAYOUTS_CONTAINER_JSON_REGEX,
+    PACKS_PRE_PROCESS_RULES_JSON_REGEX,
     PACKS_WIDGET_JSON_REGEX,
     PACKS_REPORT_JSON_REGEX,
     PACKS_RELEASE_NOTES_REGEX,
@@ -784,6 +803,7 @@ DIR_LIST_FOR_REGULAR_ENTETIES = [
     INCIDENT_TYPES_DIR,
     INCIDENT_FIELDS_DIR,
     LAYOUTS_DIR,
+    PRE_PROCESS_RULES_DIR,
     CLASSIFIERS_DIR,
     INDICATOR_TYPES_DIR,
     CONNECTIONS_DIR,
@@ -982,6 +1002,7 @@ SCHEMA_TO_REGEX = {
     'mapper': JSON_ALL_MAPPER_REGEXES,
     'layoutscontainer': JSON_ALL_LAYOUTS_CONTAINER_REGEXES,
     'layout': JSON_ALL_LAYOUT_REGEXES,
+    'pre-process-rules': JSON_ALL_PRE_PROCESS_RULES_REGEXES,
     'incidentfield': JSON_ALL_INCIDENT_FIELD_REGEXES + JSON_ALL_INDICATOR_FIELDS_REGEXES,
     'incidenttype': JSON_ALL_INCIDENT_TYPES_REGEXES,
     'image': [IMAGE_REGEX],
@@ -1239,7 +1260,8 @@ VALIDATED_PACK_ITEM_TYPES = [
     'IncidentFields',
     'IncidentTypes',
     'Classifiers',
-    'Layouts'
+    'Layouts',
+    'PreProcessRules',
 ]
 
 FIRST_FETCH = 'first_fetch'
@@ -1302,6 +1324,7 @@ class ContentItems(Enum):
     GENERIC_DEFINITIONS = 'genericdefinition'
     GENERIC_FIELDS = 'genericfield'
     GENERIC_TYPES = 'generictype'
+    PRE_PROCESS_RULES = 'pre-process-rule'
     JOB = 'job'
 
 
@@ -1320,6 +1343,7 @@ JSON_SUPPORTED_FOLDERS = {
     INCIDENT_TYPES_DIR,
     INDICATOR_FIELDS_DIR,
     LAYOUTS_DIR,
+    PRE_PROCESS_RULES_DIR,
     INDICATOR_TYPES_DIR,
     REPORTS_DIR,
     WIDGETS_DIR,
@@ -1337,6 +1361,7 @@ CONTENT_ITEMS_DISPLAY_FOLDERS = {
     REPORTS_DIR,
     INDICATOR_TYPES_DIR,
     LAYOUTS_DIR,
+    PRE_PROCESS_RULES_DIR,
     CLASSIFIERS_DIR,
     WIDGETS_DIR,
     JOBS_DIR
@@ -1356,3 +1381,7 @@ class DemistoException(Exception):
 
 
 UUID_REGEX = r'([\w]{8}-[\w]{4}-[\w]{4}-[\w]{4}-[\w]{8,12})'
+
+
+class IronBankDockers:
+    API_LINK = 'https://repo1.dso.mil/api/v4/projects/dsop%2Fopensource%2Fpalo-alto-networks%2Fdemisto%2F'