+++ demisto_sdk/commands/common/content/objects_factory.py
@@ -9,7 +9,8 @@ from demisto_sdk.commands.common.content.objects.pack_objects import (
     GenericField, GenericModule, GenericType, IncidentField, IncidentType,
     IndicatorField, IndicatorType, Integration, Job, Layout, LayoutsContainer,
     OldClassifier, OldIndicatorType, PackIgnore, PackMetaData, Playbook,
-    Readme, ReleaseNote, Report, Script, SecretIgnore, Widget)
+    PreProcessRule, Readme, ReleaseNote, ReleaseNoteConfig, Report, Script,
+    SecretIgnore, Widget)
 from demisto_sdk.commands.common.content.objects.root_objects import \
     Documentation
 from demisto_sdk.commands.common.tools import find_type
@@ -31,6 +32,7 @@ TYPE_CONVERSION_BY_FileType = {
     FileType.MAPPER: ClassifierMapper,
     FileType.LAYOUT: Layout,
     FileType.LAYOUTS_CONTAINER: LayoutsContainer,
+    FileType.PRE_PROCESS_RULES: PreProcessRule,
     FileType.REPUTATION: IndicatorType,
     FileType.INDICATOR_FIELD: IndicatorField,
     FileType.INCIDENT_FIELD: IncidentField,
@@ -40,6 +42,7 @@ TYPE_CONVERSION_BY_FileType = {
     FileType.DESCRIPTION: Readme,
     FileType.README: Readme,
     FileType.RELEASE_NOTES: ReleaseNote,
+    FileType.RELEASE_NOTES_CONFIG: ReleaseNoteConfig,
     FileType.DOC_IMAGE: DocFile,
     FileType.JAVASCRIPT_FILE: '',
     FileType.POWERSHELL_FILE: '',